n64: access compiled blocks via a hash
We'd like the recompiler to take the execution context such as kernel
mode into account when compiling blocks. That's why it's necessary to
identify blocks not just by address but all the information used at
compile time. This is done by computing a 32-bit key and using that as
a block's identifier instead of the last six physical address bits like
was done before.

The execution state and its representation as bit vector are recomputed
only when needed, in this case each time Context::setMode() is called,
which happens on powerup, in both MTC0 and MFC0 instructions, and on
exceptions.

Since we have now 32-bit instead of 6-bit keys, the block() function
hashes the keys before mapping them to one of the 64 pool rows. The hash
function was chosen arbitrarily to be better than a simple multiplicative
hash and is likely not the best choice for this exact task.